Move security modal to new component
