+++ src/components/Sidebar/AddWalletModal/AddWalletModal.module.scss
@@ -1,0 +1,3 @@
+.AddWalletModal {
+
+}
