+++ src/components/Sidebar/AddWalletModal/index.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import styles from "./AddWalletModal.module.scss";
+
+function AddWalletModal() {
+return <div className={styles.AddWalletModal}></div>;
+}
+
+export default AddWalletModal;
