+++ src/components/Sidebar/Sidebar.module.scss
@@ -59,26 +59,3 @@
     padding-right: 30px;
   }
 }
-.SecurityModal {
-  height: 300px;
-  width: 320px;
-
-  & .CheckboxContainer {
-    font-size: 14px;
-    font-weight: 500;
-
-    & .Checkbox {
-      margin-right: 5px;
-    }
-  }
-
-  & .TextboxContainer {
-    margin-top: 17px;
-  }
-
-  & .ButtonContainer {
-    display: flex;
-    justify-content: space-between;
-    margin-top: 46px;
-  }
-}
